blockcache: Reinstall LUT entries in reaper
When compiling a block that covers ours, the threaded compiler will
set the LUT entries of the various entry points. Therefore we cannot
write the LUT in lightrec_block_is_outdated(), as we would risk
overwriting the new entry points with outdated pointers.

Therefore, we now leave it to the reaper to re-install the LUT entries.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>