lightrec: Call reaper in interpreter too
When Lightrec is built with support for the threaded compiler but is
ran exclusively with the interpreter, stuff is being registered with
the reaper but lightrec_reaper_reap() is never called, so the reaper's
queue grows infinitely.

Fix this issue by calling the reaper in lightrec_run_interpreter() as
well.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>