lightrec: Pause reaper for shorter periods of time
We only need to pause the reaper for a very short period of time while
we search for the block in the block cache until we set the
BLOCK_IS_DEAD flag, otherwise the block may be removed under our feet.

As soon as the BLOCK_IS_DEAD flag is set the reaper can continue, as we
ensured that the block won't be removed by anyone but us.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>