optimizer: Don't propagate consts to REG_HI and REG_LO
lightrec_propagate_consts() uses a 32-bit bitmask of the known channels;
this means that it only encodes information for the 32 general-purpose
registers, and not for REG_HI and REG_LO.

Besides, the "values" array passed as a parameter only contains 32
values, so trying to write a value at offsets [REG_HI] and [REG_LO]
meant overwriting stack data, with the obvious noxious consequences.

Therefore, we do not want to do const propagation for these two
registers.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>