lightrec: Use default memory ops for mapped I/O registers
When lightrec_rw() is called for a memory address which corresponds to a
I/O address which can be adressed directly, use the default memory ops
instead of calling the functions provided by the glue code.

This makes the behaviour similar between the interpreter and the
dynarec, since calling into the glue code may update the target cycle.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>