lightrec: Fix use of GENMASK() in lightrec_swl/lwr
Considering the following code:

u32 foo = 32;

u32 result1 = GENMASK(31, 32);
u32 result2 = GENMASK(31, foo);

Strangely enough, result1 will resolve to 0x00000000, while result2 will
resolve to 0xffffffff. The first result may be because GCC computes the
value at compilation time without caring about overflowing the 32-bit
values.

Since GENMASK() uses a left-shift internally, the low-bit argument of
the macro must be strictly inferior than the register size. Failing to
comply with this rule causes undefined behaviour, and that's why result2
gets this value.

The lightrec_swl/lwr functions were calling GENMASK() with the low-bit
argument that can reach 32, and expected the result to be 0x0, which was
a wrong expectation as explained above.

Address this issue by using a mask of 0 if the second argument of the
GENMASK() macro would resolve to 32 in lightrec_swl/lwr.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>