emitter: Fix dynarec exit on CP0 r12/r13 write
Previously, a write on CP0 registers r12 or r13 which would trigger a
IRQ was causing lightrec_state->exit_flags to be set to
LIGHTREC_EXIT_CHECK_INTERRUPT.

However, the dispatcher loop does not check the "exit_flags" field to
know if it should exit, and only check that the cycles register is above
zero.

Some games (e.g. Jackie Chan Stuntmaster) require the IRQ to be handled
as soon as it is unmasked (or requested if it's a SWI), therefore it is
very important to exit the dynarec as soon as a IRQ is detected.

Address this issue by resetting the cycles counter if a IRQ condition is
detected, which will then force the dispatcher to exit the main loop.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>