optimizer: Detect always/never taken branches
Detect branches that are always taken or never taken using const
propagation.

Branches that are always taken can be replaced with a "beq $zero, $zero"
opcode, and branches that are never taken can be replaced with a "nop"
opcode. In that case, if the branch was local, the SYNC flag at the
target opcode can be removed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>