constprop: Handle constant propagation after branch opcodes
After the branch opcode + its delay slot, we can sometimes deduce partly
or completely the value of the branch's input registers.

For instance, with the following code:

bne t0, $zero, label
nop
addiu s0, s0, t0

We can deduce that the value of t0 will be zero at the ADDIU opcode,
because the branch would have been taken with any other value; and
therefore, in this example the ADDIU opcode is a NOP that can be
dropped.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>