emitter: Remove branch in rec_store_direct()
Improve a bit the code of rec_store_direct() by using a conditional move
instruction instead of a branch.

Because of how the code works, in the case where the target address is
not in RAM, the code LUT at offset 0x0 will be invalidated. This is
okay, because the code block at PC 0x0 shouldn't be executed that often.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>