emitter: Improve a bit rec_store_direct()
Use a mask of (RAM_SIZE - 1) in case of a 32-bit load, instead of
((RAM_SIZE - 1) & ~3), as the lower 2 bits will always be zero.

This means that the value is now a "low mask" as it extends all the way
to bit 0. The mask can then be generated in a single opcode on MIPS32r2
instead of two.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>