emitter: Generate optimized code for LWC2 and SWC2
Instead of always jumping to the C handler when a LWC2 or SWC2 opcode is
detected, generate optimized code by re-using the existing I/O and CP2
handling functions.

Since no intermediate general-purpose register is read or written in the
process, use a dedicated temporary register that will only be used in
the optimized code generated for LWC2 and SWC2.

This optimization will give a sane boost to the 3D games. For instance,
Medievil now runs about 12.5% faster on its title screen with the
skeleton.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>