optimizer: Fix find_next_reader()
It voluntarily failed to find the next opcode reading a given register
if said opcode was the delay slot of a branch, so that
lightrec_remove_useless_lui() would not try to remove it.

However, this was a problem of lightrec_remove_useless_lui() and it has
been fixed since then.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>