lightrec: Preload the block's PC into a host register
Preload the block's PC into the JIT_V0 register. This will later allow
the emitted code to generate better optimized code when dealing with
PC-relative values.

The JIT_V0 register is however not dedicated for this use, and it can be
flushed because of register pressure like any regular register.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>