lightrec: Don't trash opcode list if branches must be emulated
If some branch opcodes of the block have the "should emulate" flag, we
must call the interpreter; and therefore, the block's opcode list should
not be trashed.

This wasn't a problem until now, because a brand new block is created
just for the branch opcode; but this will change soon.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>