emitter: Jump to interpreter directly for "impossible" branches
Instead of exiting the current block and create a new one to be executed
by the interpreter when an "impossible" branch is found, jump to a
specific address in the dispatcher, which will in turn call the
interpreter.

The advantage is that the interpreter is still executing the same block;
and therefore, it can tag the branch and/or its delay slot.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>