optimizer: Handle swapped branch/DS in lightrec_reset_syncs()
The computed offset of the branch target should take into account the
fact that the branch opcode and its delay slot might have been swapped.

Until now the opcodes were never swapped when that function was called,
because of the order in which the optimization passes are performed, but
that will change in the future.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>