optimizer: Add SYNC after impossible branch
When reaching an "impossible" branch and its delay slot, the interpreter
will emulate the branch, and then will jump to the next PC - be it the
target of the branch or the address after the delay slot.

For the latter case, we want the interpreter to jump back into our
block instead of creating a new one. Add a SYNC flag to the opcode after
the delay slot so that an entry point is created.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>