interpreter: Allow impossible branches at non-zero offset
Previously, when the interpreter was called to emulate an "impossible"
branch at a non-zero offset in the current block, it would instead jump
to the PC address that corresponds to the "impossible" branch, so that
the offset was zero.

This was done so that the behaviour would match what was done by the
dynarec. However, the dynarec does not behave like that anymore, and
this mechanism can be removed.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>