lightrec: Clear LUT after "safe" branches
The opcode following a branch flagged as "impossible" and its delay slot
has a SYNC flag, so that the interpreter can re-enter the block after
the branch has been emulated. This SYNC flag might have been cleared if
runtime profiling decided that the branch was "safe". In this case, we
need to clear the code LUT at the opcode's address when re-compiling a
block, since it might not be an entry point anymore.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>