optimizer: Flag branch with loads to memory in delay slot as safe
A branch with a load opcode in its delay slot will be flagged as
"impossible", which means that the interpreter must be called to handle
it properly.

Load opcodes in delay slots of branch opcodes are problematic and tricky
to emulate, because of the load delay: if the first opcode located at
the branch's target reads the register written by the load opcode, it
will actually get the old value of the register, and not the one
resulting from the read.

As such, when Lightrec finds a branch with a load opcode in its delay
slot, it will flag the branch as "impossible", which means that the
interpreter must be called to handle it properly. However, this actually
happens quite often, and the resulting calls to the interpreter account
for quite a high percentage of Lightrec's CPU time.

One subtility of the load delays, is that in the unlikely event of an
interrupt triggering right between the load opcode and the branch
target's first opcode, then this first opcode will actually read the
new value of the register, as written by the load opcode, instead of
the old one. As such, load delays are unreliable.

As such, C compilers will never try to abuse load delays. Knowing this,
and the fact that the extreme majority of Playstation games are written
in C and compiled with GCC, only a few games actually need precise
emulation of load delays; and for the games that do, the load opcodes in
the delay slots all seem to target hardware registers and not memories
(RAM, BIOS or scratchpad).

Therefore, if runtime profiling detects that the target of the load
opcode is memory and not memory-mapped I/O registers, consider that the
branch is "safe" (aka. free of side effects), and flag it as such. This
might be a dangerous assumption, but it seems to do the trick.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>