constprop: Propagate return address of JAL(R)
Since we know the block's PC and the index of the opcode, we can deduce
the return address.

As JAL/JALR were not handled previously in the constant propagation,
their target register wasn't reset to "unknown", which caused crashes in
some games (e.g. Mortal Kombat Trilogy). This is fixed now.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>