Add "pure" and "const" function attributes when it applies
Add the "const" function attribute to the functions whose return value
depend directly on the inputs (e.g. an arithmetic function on the
parameters).

Add the "pure" function attribute to the functions whose return value
depend indirectly on the inputs (e.g. a function computing a hash of the
data pointed to by the first argument).

Signed-off-by: Paul Cercueil <paul@crapouillou.net>