optimizer: Extend load delay detection to loads outside delay slots
In the memory card handling code of OpenBIOS, there is a load opcode
whose target register is immediately used as a source register of an ALU
opcode right before. The ALU opcode effectively gets the old value of
the register and not the new one. This is something that PSX games seem
to never do, as it is not safe - an interrupt rising in-between the two
opcodes would cause the ALU instruction to obtain the new value of the
register.

Add support for it nonetheless, by simply swapping the load opcode with
the following instruction when it happens. It likely only happens in
OpenBIOS, so even though the fix is pretty unintelligent, it should
support 100% of the use cases.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>