lightrec: Add option to use the GBR as the state register on SH4
The GBR is a register reserved by the C library for handling
thread-local storage.

For the systems where TLS is not used, add an option to use the GBR
register as the state register on SH4.

Using the GBR not only frees up a regular register for Lightrec to use;
it also comes with associated memory load/store opcodes, which can
address up to PC+1020 bytes, making it ideal for loading/storing data
from/to the state structure.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>