optimizer: detect and optimize busy-wait loops
Detect the busy-wait loops that are just polling memory addresses.
Those loops don't overwrite any register that they read, so the output
after one iteration will be the exact same as after N iterations, except
for the cycle count decreasing.

For that reason above, when we find such a loop, if the branch is taken
then we know that the branch will always be taken until the next IRQ, so
we can just skip all that processing and fast-forward to the next IRQ
cycle.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>