interpreter: Add function lightrec_handle_load_delay()
This function is to be called when the first opcode at the target of a
branch or jump reads the target register of the load opcode in the
branch's delay slot.

It will run the interpreter for the first opcode at the target, then
store back the saved temporary value to the load opcode's target
register.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>