optimizer: Don't handle load delays for LWC2
The LWC2 register does not write a general-purpose register, and
therefore we don't have to handle load delays when it is found in a
delay slot.

Besides, the LWC2 opcode emitter uses the temporary register which is
also used to store the temporary value of a delayed load, which would
cause problems in games.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>