lightrec: Handle games deferencing NULL pointers
Sled Storm (and maybe others?) will access the SPU registers through a
base pointer + offset. However, it does so even before the base pointer
is initialized, resulting in a memory write to the beginning of the RAM
(aka. an undetected segmentation fault).

This causes Lightrec to tag the write opcode as hitting memory and not
the I/O address space, which then results in the I/O accesses not being
detected as such when the base pointer is eventually initialized
properly.

Address this issue by tagging the opcodes that use a base register
containing the value 0 (except $zero) as accessing I/O, which ensures
that they will be processed in C within lightrec_rw().

Fixes #14.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>