emitter: Handle MTC0 to r12 in C when running outside icache
Writing a value to the Status register (r12) can cause the instruction
cache to be isolated. This obviously cannot happen when running from a
cached window to the RAM.

The previous code assumed it only happened from BIOS code; but it can
also happen when running from a uncached window to the RAM.

This fixes ToCA World Touring Cars.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>