optimizer: Fix LWL/LWR opcodes being incorrectly swapped
Final Fantasy 8 has a code block with a LW followed by a LWL, which is
standalone - not followed by a LWR.

This triggered a false positive in the optimization pass that detects
intricated load delay dependencies, and caused the two opcodes to be
swapped.

Fix this by filtering out all the cases where the second opcode is a LWL
or LWR in the optimization pass.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>