optimizer: Fix branch opcodes being incorrectly swapped
Jumping Flash! 2 has a code block with a LW followed by a BEQ which
evaluates the load's target register.

According to the source code of PCSX, and unlike most other opcodes,
the value evaluated by the branch opcode will be the new value of the
register, and not the old one.

This should also fix Xenogears according to PCSX sources.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>