emitter: Fix "no more registers" error on 32-bit x86
32-bit x86 has a very small pool of registers, as only three
caller-saved and three callee-saved registers are available.

From those six registers, two are reserved by Lightrec (one for the
state pointer, one for the cycle counter), which means that at any given
time, only four registers can be requested at the same time.

rec_store_memory() uses a maximum of four registers; however, when a
store-to-memory opcode is in the delay slot of a branch or jump, an
extra register containing the next PC value will be requested and
locked, which means that the maximum number of registers that can be
used at any given time is actually three.

Thanksfully, the rec_store_memory() function can be updated to use only
three registers (on little-endian systems), by freeing the "rs" register
as soon as possible.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>