emitter: Reset register cache after jump to EOB
Sometimes the block will continue after an unconditional jump to the
end-of-block handler, for instance after a BREAK opcode.

In that case, it's better to reset the register cache, because the
JIT_V0 host register is otherwise locked to the next PC value and not
available anymore for the dynarec.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>