optimizer: Add unsafe optimization that assumes $sp/$gp always hit RAM
Lightrec makes the assumption that the $sp and $gp registers always
point to memory, be it RAM, BIOS or the scratchpad. This assumption can
obviously be broken with custom code, but has been proven true with all
regular PSX games.

Add an unsafe optimization that will assume that the $sp and $gp
registers always point to RAM and nothing else. This will make the
generated code faster, but not all PSX games will support it.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>