emitter: Optimize register allocation
The instructions that use one or two input registers and one output
register would make the opcode generators reserve between two and three
host registers: one for each input and one for the output.

However, if the first input register is to be unloaded or discarded
after the operation, then it is possible to use the same host register
for both this register and the output register; the host register will
point to the input register before the instruction, and point to the
output register after the instruction.

This generates slightly better code, for two reasons. The first one, is
that there is less register pressure, which will reduce the number of
register spill/reloads. The second one, is that some archs will generate
better code when the first input register and the destination register
of the underlying host instruction are the same.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>