optimizer: Tag LUI + {ORI,ADDI,ADDIU} as a 32-bit immediate load
When a LUI is followed by a ORI, ADDI or ADDIU and the source and
destination registers match the one used in the LUI, this is effectively
a 32-bit immediate load to this register.

Tag the LUI opcode and the follow-up ORI/ADDI/ADDIU, so that the code
emitter can generate a 32-bit immediate load at once.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>