emitter: Optimize access to PSX timers
Instead of performing a division every time we want to compute the value
of a timer, pre-compute M/P values in lightrec_set_timer_data().

These values can then be used to compute the value of a timer using a
high-multiply operation, which will be much faster than performing a
division.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>