lightrec: Use ANSI C types for fixed-size typedefs
Use ANSI C types instead of C99 <stdint.h> types, as the latter are
annoying to use: they may be defined differently on different archs
(e.g. uint32_t is generally "unsigned int", but for some reason is
"unsigned long" on SH-4), and require specific format tokens for the
printf() family of functions.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>