lightrec: Pass full register value to store callbacks
It appears the r3000 puts the whole register on bus regardless if you
write byte, half or full word. It's up to the receiving device to
figure out what kind of write it is from the control signals. But
apparently some hw registers don't look at the control signals at all
and just latch the whole bus.

Support this case by passing the full 32-bit value of the register to
store callbacks.

Fixes #26.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>