lightrec: Fix reap of covered blocks
If we overwrite the code LUT and reap the underlying block of each
branch target of the current block being recompiled, we have a race,
because the underlying blocks may be used by different entry points in
the code LUT that we still haven't cleared.

The only solution is to first mark all blocks as dead, and make sure
they won't be recompiled, and then clear them from the code LUT all at
once.

This should fix the (hopefully) last race of the threaded compiler.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>