interpreter: Fix RFE to GTE opcode
I am not sure why the previous code checked that the interrupt bits in
the Cause register were all cleared; what I do know is that it was not
working properly, and caused crashes in Um Jammer Lammy and graphical
glitches in Vagrant Story and Colin Mc Rae Rally.

Handle RFE to GTE opcodes by actually checking that the EPC register
points to a GTE opcode.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>