interpreter: Fix fully tagged blocks being recompiled
The dispatcher would call into the interpreter to handle load delay
slots. The interpreter would load the opcode at the branch target from
RAM, and initialize a 'struct opcode' with the flags set to 0x0.

This caused lightrec_rw() to try to tag the opcode, and request a
recompilation of the block, even if the block was fully compiled.

This fixes crashes in Formula One Arcade.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>