lightrec: Always unregister blocks in reaper
When the threaded compiler is enabled, always unregister blocks from the
block cache in the reaper. Otherwise, there is a race between the main
thread running lightrec_get_block() and the threaded compiler, after it
gets a handle from the block cache, and before it checks for its status
flags.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>