lightrec: Only read block cache once when compiling block
Reading it twice is a race vs. the main thread. The scenario is, when
compiling block 1, covering blocks 2 and 3, those last two are marked
as dead. Meanwhile, the main thread wants to execute block 2, sees that
it is marked as dead, and will re-create a brand new block 2 at the same
block LUT address. When the threaded compiler re-loads all the blocks to
destroy them (with the reaper), it will then get a pointer to the new
block 2, instead of a pointer to the old block 2.

Address this issue by using a huge local array that will be filled with
pointers to the blocks covered by the block being compiled; then the
correct blocks will be destroyed even if their respective entries in the
code LUT now point to freshly re-generated blocks.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>