recompiler: Introduce a new priority system
Each block added to the recompiler queue now has a counter, originally
set to 1 when the block is first added to the queue, that will be
incremented each time the main thread requests the block to be
compiled.

This ensures that the blocks that are needed the most will be compiled
in priority, instead of giving priority to new blocks yet to be
compiled, which may actually never be executed again.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>