optimizer: Improve transformation of LUI opcodes
The code was checking that the LUI to modify is not a delay slot, by
checking that the previous opcode is not a branch. This won't match when
the previous opcode is a branch that was swapped with its delay slot.

Use the is_delay_slot() function instead, which performs the proper
checks.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>