Remove support for different memory offsets
Supporting different memory offsets for the RAM, BIOS, I/O etc. makes
the code much more complicated than it should be, and makes it even
harder to improve it further.

Therefore, Lightrec now requires MMU support. This is fine as it would
be pretty slow on platforms without a MMU anyway.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>