lightrec: Free all blocks when changing ratio of cycles per opcode
If the ratio of cycles per opcode has changed, all the previously
compiled blocks are now invalid, and have to be dropped.

Pause the threaded recompiler (so that it won't try to compile blocks
we're about to remove), reap any pending data, then free all blocks and
unpause the threaded compiler.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>