emitter: Remove counter-productive optimization for MOV meta-opcode
Previously, the generator code for the MOV meta-opcode checked that if
the source register was not loaded, its value would be loaded directly
into the target register.

However, this was counter-productive, as at that point of the code we
know that the source register will eventually be used (as the opcode
does not have the "unload rs" or "discard rs" flags). Loading its value
directly into the target register meant that the value would eventually
be loaded once again from the register cache into the source register.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>