optimizer: Do not swap load opcodes when next op is a branch
There is an actual one-instruction delay between the moment a load
opcode is decoded, and the moment the new value is available in the
target register; and therefore, if the very next instruction reads the
target register, it will read the old value of the register, not the new
one.

This setup is actually very rare, because if an interrupt fires at the
right time, the second opcode will read the new value of the register,
making this setup unreliable.

To support OpenBIOS which uses this setup in its memory card code,
Lightrec will attempt to swap the load opcode with the next opcode.

This obviously does not work when the second opcode is a branch. In that
case, by lack of a better way to handle it, it's better to just ignore
the setup completely, as it does more harm than good.

This fixes a crash in "Goemon Shin Sedai Shuumei".

Signed-off-by: Paul Cercueil <paul@crapouillou.net>