lightrec: Introduce preliminary support for external code LUT
Introduce the ENABLE_EXTERNAL_CODE_LUT option in CMake.

When enabled, and in the case where the emulator populates the
PSX_MAP_CODE_LUT in the memory map and the memory map itself makes it
possible, the external LUT will be used.

What that means, is that instead of allocating about 2.5 MiB for the
code LUT (2 MiB for the emulated RAM + 512 KiB for the BIOS), Lightrec
will use the given memory address as a pointer to the code LUT.

The pointer itself has to point to a memory area which verifies each of
the following conditions:
- (pointer + 0x0), (pointer + 0x20'0000), (pointer + 0x40'0000) and
  (pointer + 0x60'0000) are memory windows to the same 2 MiB buffer;
- (pointer + 0x1fc0'0000) is a memory mapping to a 512 KiB buffer;
- (pointer + 0x1f80'0000) is a memory mapping to a 1 KiB buffer;
- (pointer + 0x1f80'03ff) must be a 32-bit address (which basically
  means that everything between the pointer and the end of the 1 KiB
  buffer must be in 32-bit memory space).

Using this setup speeds up function lookup and code invalidation, as the
kunseg'd addresses can be used directly as offsets from the external
code LUT.

Since the emitter has to be patched as well, the code will currently not
set state->external_lut. It will be enabled in a future commit when
everything is in place, in order to keep the code bisectable.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>