recompiler: Improve latency on single-core systems
On single-core CPUs, if we got a request for a block that's already in
the queue, we'll probably get many more before the compiler thread can
run, which means that the block will be interpreted until then,
wasting a lot of performance. In that case, it is better to just let
the compiler thread run now.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>